% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epheremis.R
\name{ephemeris}
\alias{ephemeris}
\title{Get ephemeris data}
\usage{
ephemeris(
  name = "s:Sun",
  longitude = 0,
  latitude = 0,
  t0 = Sys.Date(),
  nbd = 5,
  step = 1,
  debug = FALSE
)
}
\arguments{
\item{name}{character value of object in question, e.g. \code{"s:Sun"} for the sun and
\code{"s:Moon"} for the moon.  Default: \code{"s:Sun"}.}

\item{longitude}{decimal value of the longitude. Default: 0.}

\item{latitude}{decimal value of the latitude.  Default: 0.}

\item{t0}{POSIXct time (UTC) for the first retrieval, or a string from which
\code{\link[=as.POSIXct]{as.POSIXct()}} can infer such a time. Default: start of present day.}

\item{nbd}{integer value of the number of times to be retrieved. Default: 5.}

\item{step}{numeric value for the time increment, in days.  Default: 1.}

\item{debug}{logical value indicating whether to print debugging information. At the moment,
the only information printed is the query string. Default: \code{FALSE}.}
}
\value{
\code{ephemeris} returns a data frame containing columns named
\code{"time"}, \code{"RA"} (string hour angle), \code{"RAdec"} (decimal hour angle),
\code{"DEC"} (string hour angle), \code{"DECdec"} (decimal hour angle),
\code{"Dobs"}, \code{"VMag"}, \code{"Phase"}, \code{"Elong."},
\code{"dRAcosDEC"}, \code{"dDEC"}, and \code{"RV"}.  Note that the numeric values
\code{RAdec} and \code{DECdec} are computed from the string values \code{RA} and \code{DEC} retrieved
from the server.
}
\description{
This function infers ephemeris data by querying a website (reference 1),
and so it requires a web connection to work. It is aimed at people who
already know the ideas and terminology of ephemeris computations.
}
\details{
Apart from the parameters of this function, the other specifications for
the query are set up in the same way as is used in the query-generation examples.
For example, the 'observer' is set to \verb{@500}, designating the centre of the earth.
Users who are curious about the query should specify \code{debug=TRUE} when calling
this function, and they are encouraged to contact the package author, if
they would like any of these hard-wired defaults to be transformed
into a user-adjustable value, via the creation of new parameters
to this function.
}
\examples{
# Plot daily Right Ascension and Declination values over a 28-day period.
library(ephemeris)
s <- ephemeris(nbd=28)
m <- ephemeris("s:Moon", nbd=28)
par(mfrow=c(2, 1), mar=c(3,3,1,2), mgp=c(2,0.7,0))
RAlim <- range(c(s$RAdeg, m$RAdeg))
plot(s$time, s$RAdeg, type="o", xlab="", ylab="RA [hour angle]", col=2, ylim=RAlim)
lines(m$time, m$RAdeg, col=4, type="o")
mtext("Red: sun", col=2, adj=0)
mtext("Blue: moon", col=4, adj=1)
DEClim <- range(c(s$DECdeg, m$DECdeg))
plot(s$time, s$DECdeg, type="o", xlab="", ylab="DEC [hour angle]", col=2, ylim=DEClim)
lines(m$time, m$DECdeg, col=4, type="o")

}
\references{
\enumerate{
\item The \emph{Institut de mécanique céleste et de calcul des éphémérides} website
\url{https://ssp.imcce.fr/webservices/miriade/api/ephemcc/} is the source of
data returned by \code{ephemeris}.
}
}
\author{
Dan Kelley
}
