% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephemeris.R
\name{ephemeris}
\alias{ephemeris}
\title{Get ephemeris data}
\usage{
ephemeris(
  name = "p:Sun",
  longitude = 0,
  latitude = 0,
  t0 = Sys.Date(),
  nbd = 5,
  step = 1,
  theory = "INPOP",
  teph = 1,
  tcoor = 1,
  rplane = 1,
  debug = FALSE
)
}
\arguments{
\item{name}{character value of object in question. There are many objects
that can be checked, e.g. \code{"p:Sun"} for the sun and
\code{"s:Moon"} for the moon.  Default: \code{"p:Sun"}.}

\item{longitude}{decimal value of the longitude. Default: 0.}

\item{latitude}{decimal value of the latitude.  Default: 0.}

\item{t0}{POSIXct time (UTC) for the first retrieval, or a string from which
\code{\link[=as.POSIXct]{as.POSIXct()}} can infer such a time. Default: start of present day.
In Reference 1, this is called \code{ep} and referred to as the requested epoch.}

\item{nbd}{integer value of the number of times to be retrieved. Default: 5.}

\item{step}{numeric value for the time increment, in days.  Default: 1.}

\item{theory}{character value indicating the planetary theory, with choices
\code{"INPOP"}, \code{"DE200"}, \code{"BDL82"}, \code{"SLP98"}, \code{"DE403"},
\code{"DE405"}, \code{"DE406"}, \code{"DE430"}, and \code{"DE431"}. Default: \code{"INPOP"}.}

\item{teph}{integer value for the type of ephemeris, with 1
for astrometric J2000, 2 for apparent of the date, 3 for mean
of the date, or 4 for mean J2000. Default: 1.}

\item{tcoor}{integer value for the type of coordinate, with 1
for spherical, 2 for rectangular, 3 for local, 4 for hour angle, and
5 dedicated to observation. Default: 1.}

\item{rplane}{integer value for the reference plane, with 1
for equator and 2 for ecliptic. Default: 1.}

\item{debug}{logical value indicating whether to print debugging information. At the moment,
the only information printed is the query string. Default: \code{FALSE}.}
}
\value{
\code{ephemeris} returns a data frame containing columns named
\code{"time"}, \code{"RA"}, \code{"RAdec"} (decimal hour angle),
\code{"DEC"} (string hour angle), \code{"DECdec"} (decimal hour angle),
\code{"Dobs"}, \code{"VMag"}, \code{"Phase"}, \code{"Elong."},
\code{"dRAcosDEC"}, \code{"dDEC"}, and \code{"RV"}.  In addition, \emph{PROVISIONALLY}
decimal values of \code{RA} and \code{DEC} are computed if \code{tcoor=1}, but
these may be incorrect.
}
\description{
This function infers ephemeris data by querying a website (reference 1),
and so it requires a web connection to work. It is aimed at people who
already know the ideas and terminology of ephemeris computations. Tests
against the \emph{In the Sky} website suggest RA and DEC accuracy to within
1 arc-second for the sun, and 15 arc-seconds for the moon.
}
\details{
The names of parameters, and their explanaations, are patterned on Reference 1.
In particular, the moon is \code{"s:Moon"} and the sun is (oddly) \code{"p:Sun"}.

Apart from the listed parameters of this function, the other specifications for
the query are set up in the same way as is used in the query-generation examples
provided in Reference 1.
For example, the 'observer' is set to \verb{@500}, designating the centre of the earth.

Users who are curious about the query should specify \code{debug=TRUE} when calling
this function, and they are encouraged to contact the package author, if
they would like any of these hard-wired defaults to be transformed
into a user-adjustable value, via the creation of new parameters
to this function.
}
\examples{
# Plot daily Right Ascension and Declination values over a 28-day period.
library(ephemeris)
s <- ephemeris("p:Sun", nbd=28)
m <- ephemeris("s:Moon", nbd=28)
par(mfrow=c(2, 1), mar=c(3,3,1,2), mgp=c(2,0.7,0))
RAlim <- range(c(s$RAdec, m$RAdec))
plot(s$time, s$RAdec, type="o", xlab="", ylab="Right Ascension [hour]", col=2, ylim=RAlim)
lines(m$time, m$RAdec, col=4, type="o")
mtext("Red: sun", col=2, adj=0)
mtext("Blue: moon", col=4, adj=1)
DEClim <- range(c(s$DECdec, m$DECdec))
plot(s$time, s$DECdec, type="o", xlab="", ylab="Declination [angle]", col=2, ylim=DEClim)
lines(m$time, m$DECdec, col=4, type="o")

}
\references{
\enumerate{
\item The \emph{Institut de mécanique céleste et de calcul des éphémérides} website
\url{https://ssp.imcce.fr/webservices/miriade/api/ephemcc/} is the source of
data returned by \code{ephemeris}.
\item The \emph{In The Sky} website used for tests demonstrating 1 arc-second
consistency for sun RA and DEC, and 15 arc-second consistency for moon:
\url{https://in-the-sky.org/ephemeris.php}
}
}
\author{
Dan Kelley
}
