% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epheremis.R
\name{ephemeris}
\alias{ephemeris}
\title{Get ephemeris data from ssp.imcce.fr}
\usage{
ephemeris(
  name = "s:Sun",
  longitude = 0,
  latitude = 0,
  t0 = Sys.Date(),
  nbd = 5,
  step = 1,
  debug = FALSE
)
}
\arguments{
\item{name}{character value of object in question, e.g. \code{"s:Sun"} for the sun and
\code{"s:Moon"} for the moon.  Default: \code{"s:Sun"}.}

\item{longitude}{decimal value of the longitude. Default: 0.}

\item{latitude}{decimal value of the latitude.  Default: 0.}

\item{t0}{POSIXct time (UTC) for the first retrieval, or a string from which
\code{\link[=as.POSIXct]{as.POSIXct()}} can infer such a time. Default: start of present day.}

\item{nbd}{integer value of the number of times to be retrieved. Default: 5.}

\item{step}{numeric value for the time increment, in days.  Default: 1.}

\item{debug}{logical value indicating whether to print debugging information. At the moment,
the only information printed is the query string. Default: \code{FALSE}.}
}
\value{
a data frame containing columna named
\code{"time"}, \code{"RA"}, \code{"RAdeg"}, \code{"DEC"}, \code{"DECdeg"}, \code{"Dobs"}, \code{"VMag"}, \code{"Phase"}, \code{"Elong."},
\code{"dRAcosDEC"}, \code{"dDEC"}, and \code{"RV"}.  All items except those ending in \code{deg} are as is
returned form the sever.  However, since the angles \code{"RA"} and \code{"DEC"} are in a non-numeric
format, \code{ephemeris} computes degimal forms, stored in \code{"RAdeg"} and \verb{DECdeg"}.
}
\description{
\code{ephemeris} infers the ephemeris data by querying
\url{https://ssp.imcce.fr/webservices/miriade/api/ephemcc/}, and so
a working web connection is required for it to work.
}
\examples{
# Plot daily Right Ascension and Declination values over a 28-day period.
library(ephemeris)
s <- ephemeris(nbd=28)
m <- ephemeris("s:Moon", nbd=28)
par(mfrow=c(2, 1), mar=c(3,3,1,2), mgp=c(2,0.7,0))
RAlim <- range(c(s$RAdeg, m$RAdeg))
plot(s$time, s$RAdeg, type="o", xlab="", ylab="Right Ascension", col=2, ylim=RAlim)
lines(m$time, m$RAdeg, col=4, type="o")
mtext("Red: sun", col=2, adj=0)
mtext("Blue: moon", col=4, adj=1)
DEClim <- range(c(s$DECdeg, m$DECdeg))
plot(s$time, s$DECdeg, type="o", xlab="", ylab="Declination", col=2, ylim=DEClim)
lines(m$time, m$DECdeg, col=4, type="o")

}
\references{
\url{https://ssp.imcce.fr/webservices/miriade/api/ephemcc/}
}
\author{
Dan Kelley
}
